pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.12.0",
    "jupyterlab>=4.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.25",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openrouter>=0.2.1",
    "langgraph>=1.1.5",
    "python-dotenv>=1.2.2",
]
